fix: Main Page uses server side loading to prevent jitter
