+++ www/middleware.ts
@@ -2,10 +2,7 @@ import { type NextRequest } from "next/server";
 import { updateSession } from "@/utils/supabase/middleware";
 
 export async function middleware(request: NextRequest) {
-  console.log('Middleware - Request method:', request.method);
-  const response = await updateSession(request);
-  console.log('Middleware - Response status:', response.status);
-  return response;
+  return await updateSession(request);
 }
 
 export const config = {
