+++ www/package.json
@@ -11,7 +11,7 @@
     "sync-stripe": "node scripts/stripeSync.mjs"
   },
   "dependencies": {
-    "@openrouter/ai-sdk-provider": "^0.0.6",
+    "@ai-sdk/openai": "^1.0.6",
     "@radix-ui/react-label": "^2.1.0",
     "@radix-ui/react-slot": "^1.1.0",
     "@radix-ui/react-tabs": "^1.1.1",
